Tugas 6 selesai

+++ main/views.py
@@ -1,6 +1,6 @@
 import datetime
 from django.shortcuts import get_object_or_404, render, redirect
-from django.http import HttpResponseRedirect, HttpResponse
+from django.http import HttpResponseNotFound, HttpResponseRedirect, HttpResponse
 from main.forms import ItemForm, Item
 from django.urls import reverse
 from django.core import serializers
@@ -8,6 +8,7 @@ from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages 
 from django.contrib.auth.decorators import login_required
+from django.views.decorators.csrf import csrf_exempt
 
 
 # Create your views here.
@@ -89,28 +90,55 @@ def logout_user(request):
     response.delete_cookie('last_login')
     return response
     
+@csrf_exempt
 def add_amount(request, item_id):
-    item = get_object_or_404(Item, pk=item_id) #Mengakses item yang ingin dimodifikasi
-    item.user = request.user; 
-    if item.amount > 0:
-        item.amount += 1
-        item.save()
-    return redirect('main:show_main')
+    if request.method == 'POST':
+        item = Item.objects.get(pk=item_id) #Mengakses item yang ingin dimodifikasi
+        item.user = request.user
+        if item.amount > 0:
+            item.amount += 1
+            item.save()
+        return HttpResponse(b"ADDED", status=201)
+    return HttpResponseNotFound()
 
+@csrf_exempt
 def reduce_amount(request, item_id):
-    item = get_object_or_404(Item, pk=item_id)
-    item.user = request.user
-    if item.amount > 1:
-        item.amount -= 1
-        item.save()
-    else:
-        item.delete()
-    return redirect('main:show_main')
+    if request.method == 'POST':
+        item = Item.objects.get(pk=item_id)
+        item.user = request.user
+        if item.amount > 1:
+            item.amount -= 1
+            item.save()
+        else:
+            item.delete()
+        return HttpResponse(b"REDUCED", status=201)
+    return HttpResponseNotFound()
 
-def delete_item(request, id):
-    # Get data berdasarkan ID
-    item = Item.objects.get(pk = id)
-    # Hapus data
-    item.delete()
-    # Kembali ke halaman awal
-    return HttpResponseRedirect(reverse('main:show_main'))
+@csrf_exempt
+def delete_item(request, item_id):
+    if request.method == 'DELETE':
+        item = Item.objects.get(pk=item_id, user=request.user)
+        item.user = request.user
+        item.delete()
+        return HttpResponse(b"REMOVED", status=201)
+    return HttpResponseNotFound()
+
+def get_item_json(request):
+    item = Item.objects.filter(user=request.user)
+    item.user = request.user
+    return HttpResponse(serializers.serialize('json', item))
+
+@csrf_exempt
+def add_item_ajax(request):
+    if request.method == 'POST':
+        name = request.POST.get("name")
+        price = request.POST.get("price")
+        amount = request.POST.get("amount")
+        description = request.POST.get("description")
+        user = request.user
+
+        new_product = Item(name=name, price=price, description=description, amount=amount, user=user)
+        new_product.save()
+
+        return HttpResponse(b"CREATED", status=201)
+    return HttpResponseNotFound()
